# modules/rest-api/src/com/haulmont/addon/restapi/api/rest-api-swagger.yaml
swagger: '2.0'
info:
  version: "7.3"
  title: CUBA Platform REST API
  description: |

    ## General Overview

    CUBA Platform REST API provides the following functionality:

    * Entities CRUD operations
    * Execution of predefined JPQL queries
    * Execution of services methods
    * Getting metadata (entities, views, enumerations, datatypes)
    * Getting localized messages
    * Getting current user permissions (access to entities, attributes, specific permissions)
    * Getting current user information (name, language, time zone, etc.)
    * Working with files

    The REST API is available at the URL `/rest/v2` both in the web and the portal modules, e.g. `http://localhost:8080/app/rest/v2/entities/sales$Order`

    ## Authentication

    ### Authentication using the OAuth2 protocol

    CUBA Platform REST API uses an OAuth2 resource owner password credentials grant to protect the endpoints (https://tools.ietf.org/html/rfc6749#page-37).

    Before accessing a protected resource, a user must get an authentication token. The token can be obtained at the `/oauth/token` URL.

    The `/oauth/token` endpoint is protected with the basic authentication. The request must contain an `Authorization` header for basic
    authentication with REST API client credentials. Please note that client credentials are not the same as platform user credentials.
    By client we mean an application that uses the REST API. The default client credentials are defined by the following
...

...

...
basePath: /app/rest/v2
produces:
...
consumes:
...
definitions:
...
  fileInfo:
    type: object
    properties:
      id:
        type: string
        example: c994111a-cdfa-9695-95fc-bbbb63a504a7
      name:
        type: string
        example: sample-file.docx
      size:
        type: number
        example: 105396
...
securityDefinitions:
  oauth2:
    type: oauth2
    flow: password
    tokenUrl: http://localhost:8080/app/rest/v2/oauth/token
    scopes:
      rest-api: use REST API
  tokenEndpointBasic:
    type: basic
security:
  - oauth2:
      - rest-api

################################################################################
#                                   Paths                                      #
################################################################################
paths:
...
  /files:
    post:
      tags:
        - Files
      summary: Upload a file
      description: |
        Uploads a file. The method behavior may be different depending of the 'Content-Type' header value.

        If the header value is 'multipart/form-data' then file contents must be put to the part called 'file'.

        If the 'Content-Type' header has some other value then file contents must be passed in the request body.
        The 'Content-Length' header will be used for FileDescriptor 'size' field value.

        File upload may be disabled for the security role using the `cuba.restApi.fileUpload.enabled` specific permission.
      parameters:
        - name: name
          in: query
          type: string
          required: false
          description: |
            The file name. If this parameter is missing then a FileDescriptor id will be used as a file name. In case of
            `multipart\form-data` request a file name may be taken from the `filename` attribute of the `Content-Disposition`
            header.
          default: ""
      responses:
        201:
          description: |
            File uploaded. FileDescriptor information is returned in the response body.
          schema:
            $ref: '#/definitions/fileInfo'
          headers:
            location:
              type: string
              description:
                A URL the uploaded file can be downloaded with
  /files/{id}:
    get:
      tags:
        - Files
      summary: Download the file
      description: Downloads the file
      parameters:
        - name: id
          in: path
          type: string
          description: |
            FileDescriptor identifier
          default: ""
          required: true
        - name: attachment
          in: query
          type: boolean
          required: false
          description: |
            If set to 'true' then the 'Content-Disposition' header of the response will be set to 'attachment'. 'inline' value
            will be used otherwise.
          default: ""
      responses:
        200:
          description: Successfully downloaded. The file contents is returned in the response body.
        403:
          description: Forbidden. The user doesn't have permissions to read the 'sys$FileDescriptor' entity.
        404:
          description: File with the given ID not found


############################### Metadata #######################################
...